Structure reading in image into class
Screw performance concerns, i want neatness!
Long term it might be worth looking into how we could structure the
memory the image readin gives us to be neatly accessible (or maybe
just wrap the nasty (line_idx*width*3)+(pixel_idx*3)+color_idx
calculation in there? We could obviously free the memory in the
destructor, and I'm not even sure we need to keep it around the
whole time, since Ideally we'll abstract these values before
clustering or whatever.

I've rudimentarily tested the correctness of how we're using the
stb-function by comparing the value of the bottom right pixel to the
value GIMP tells me for it in a reference image:
  Pixel bottom_right = img.at(img.get_height()-1,img.get_width()-1);
  printf("bottom right pixel has values (%d,%d,%d)\n",
        bottom_right.r,
        bottom_right.g,
        bottom_right.b
      );
This turned up (3,255,138) just as GIMP told me, so we know we don't
have our RGBs twisted around or have any skew where it gives us 4
values instead of the 3 we ask for (because this ref img was a png
with 4 channels).
Seems to be fine.