Allow for threshold setting
Not interested in sketching out a full CLI already, it would be the
case that what we're doing now is the 'naive' mode or something
along those lines, in which you can set a threshold.
For now, if you give 2 params, theyre treated as
  <threshold> <image>
just one is treated as
  <image>
and none is an error, because we need an image.