Write a simple cli parser and use it
I was honestly just tired of the parsing libraries on offer.
getopt is obviously solid, but I'm not looking to toss C-strings
around.
boost::program_options is very poorly documented and didn't make
much sense to me in its design.
I'd used cxxopts in the past and the design makes sense to me, but
within a few minutes of experimenting I'd found the unexpected
behaviour, that if I specified an implicit value for an option, then
that option would simply ignore a value passed in the
`--option=value` syntax and only take it via `--option value`.
Even though I'll probably have bigger gaps in my implementation, its
just tiresome to dig into these complex libraries for something as
simple as parsing arguments and having additional dependencies, when
their value add is so minimal here.