Avoid custom iterator relying on undefined behavior
Right now (and probably not for much longer, since I realized that
contiguous memory would actually be pretty nice to have here) we
have a `vector<vector<Pixel>>` here, and thus need to keep track of
two iterators (one v<v<P>>::iterator, one v<P>::iterator).
This meant it was easiest to model the end() iterator simply as
`two_dim_data.end()->begin()`, but according to <cppreference.com>
accessing `end()` (in this case by dereferencing it through `->`) is
undefined behavior.
Instead of doing that here we now have two separate constructors for
our Iterator and pass them the final end() iterator as a reference
to compare to, so they can avoid dereferencing it in `operator++`.

This is pretty obviously hacky and the iterator really doesn't give
us to much anyways, I was mostly curious how to write it and it
turned out more complicated than I thought.
For a contiguous block of memory, whether managed by some Vec2D
class from some library or by some container we roll ourselves, it
would of course be a lot simpler but I wanted to commit this state
for posterity.